# openspec/openapi-v1.yaml
openapi: 3.1.0
info:
  title: AstrBot OpenAPI v1
  version: 0.1.0
  description: >
    Target REST contract for migrating AstrBot HTTP APIs to /api/v1.
    Dynamic AstrBot configuration payloads are intentionally modeled as open
    JSON objects because their schemas are provided at runtime by template
    endpoints.
servers:
  - url: http://localhost:6185/api/v1
    description: Local AstrBot server
security:
  - ApiKeyAuth: []
...
paths:
  /auth/login:
    post:
      tags: [Auth]
      summary: Login to the dashboard API
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
        "401":
          $ref: "#/components/responses/Error"
...
  /auth/setup:
    post:
      tags: [Auth]
      summary: Complete first-run dashboard account setup
      operationId: setupAuth
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetupAuthRequest"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /auth/totp/setup:
    post:
      tags: [Auth]
      summary: Start or refresh TOTP setup
      operationId: setupTotp
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TotpSetupRequest"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
...
  /auth/account:
    patch:
      tags: [Auth]
      summary: Update dashboard account credentials
      operationId: updateAuthAccount
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAccountRequest"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api-keys:
    get:
      tags: [API Keys]
      summary: List OpenAPI keys
      operationId: listApiKeys
      responses:
        "200":
          $ref: "#/components/responses/Ok"
    post:
      tags: [API Keys]
      summary: Create an OpenAPI key
      operationId: createApiKey
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateApiKeyRequest"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api-keys/{key_id}/revoke:
    post:
      tags: [API Keys]
      summary: Revoke an OpenAPI key
      operationId: revokeApiKey
      parameters:
        - $ref: "#/components/parameters/KeyId"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api-keys/{key_id}:
    delete:
      tags: [API Keys]
      summary: Delete an OpenAPI key
      operationId: deleteApiKey
      parameters:
        - $ref: "#/components/parameters/KeyId"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Ok:
      description: Standard AstrBot success response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/SuccessEnvelope"
    Error:
      description: Standard AstrBot error response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
  parameters:
...
    KeyId:
      name: key_id
      in: path
      required: true
      schema:
        type: string
...
  schemas:
    SuccessEnvelope:
      type: object
      required: [status, data]
      properties:
        status:
          type: string
          const: ok
        message:
          type: string
        data: {}
      additionalProperties: false
    ErrorEnvelope:
      type: object
      required: [status, message]
      properties:
        status:
          type: string
          const: error
        message:
          type: string
        data: {}
      additionalProperties: false
...
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
        code:
          type: string
          description: TOTP code or recovery code when two-factor authentication is required.
        trust_device_flag:
          type: boolean
      additionalProperties: false
    SetupAuthRequest:
      type: object
      required: [username, password, confirm_password]
      properties:
        username:
          type: string
        password:
          type: string
        confirm_password:
          type: string
      additionalProperties: false
    UpdateAccountRequest:
      type: object
      required: [password]
      properties:
        password:
          type: string
        new_password:
          type: string
        confirm_password:
          type: string
        new_username:
          type: string
      additionalProperties: false
    TotpSetupRequest:
      type: object
      properties:
        code:
          type: string
        secret:
          type: string
      additionalProperties: false
    CreateApiKeyRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        scopes:
          type: array
          items:
            type: string
          example: [chat, config, file, im]
        expires_at:
          type: string
          format: date-time
        expires_in_days:
          type: integer
          minimum: 1
      additionalProperties: false
...